+++ src/constants.ts
@@ -10,6 +10,8 @@ export const TASK_SAVE_SUCCESS_MESSAGE = 'New task was created';
 export const TASK_UPDATE_MESSAGE = 'Task updated';
 export const TASK_UPDATE_ERROR_MESSAGE = 'An error occurred while updating';
 
+export const DELETE_SPACE_MESSAGE = "Are you sure you want to eliminate this Space?";
+
 export const DEFAULT_TASK_DETAILS = [
     'id', 'name', 'description', 'url', 'priority', 'creator', 'tags'
 ];