+++ src/extension.ts
@@ -7,9 +7,11 @@ import { tokenService } from './token/service';
 import { EditWebview } from './web_view/edit/editWebview';
 import { MainProvider } from './tree_view/main_provider';
 import { NewTaskWebview } from './web_view/new/newTaskWebview';
+import { Utils } from './utils';
 
 
 export async function activate(context: vscode.ExtensionContext) {
+	var utils = new Utils(vscode.window);
 	let storageManager = new LocalStorageService(context.workspaceState);
 	tokenService.init(storageManager);
 	var token: any = await storageManager.getValue('token');
@@ -59,6 +61,23 @@ export async function activate(context: vscode.ExtensionContext) {
 					vscode.window.showInformationMessage('good, your task is safe');
 				}
 			});
+	});
+
+	vscode.commands.registerCommand('clickup.addSpace', () => {
+		vscode.window.showInputBox({
+			prompt: "Insert space name"
+		}).then((response) => {
+			console.log('response', response);
+		});
+	});
+
+	vscode.commands.registerCommand("clickup.deleteSpace", (spaceItem) => {
+		console.log('space', spaceItem.id);
+		utils.confirmDialog(constants.DELETE_SPACE_MESSAGE, async () => {
+			var response = await wrapper.deleteSpace(spaceItem.id);
+			console.log('response', response);
+			// provider.refresh();
+		});
 	});